221111 이미지 휠로 배율 변화 구현
- 픽쳐컨트롤 좌표에 대한 감은 잡았으니 이제 스크롤바 버튼 구현 후 드래그 드랍으로 진행.

- 내 생각으론 배율을 곱한 커진 값에서 원본 사이즈를 빼는 식으로 계산이 되어야 하는데 가감이나 승제 개념이 반대인 이유를 모르겠음. 그것 때문에 좀 헤맸음.

+++ File/ViewerTemp/ViewerTemp/ViewerTempDlg.h
@@ -13,11 +13,10 @@ public:
 	CPoint m_ptMouse;
 	CRect Rect;
 	float m_pos = 1.f;
-	double origin_w, origin_h;
+	int origin_w, origin_h;
 	double rect_w_center, rect_h_center;
 	double loc_x, loc_y;
 	double show_w, show_h;
-
 
 // 대화 상자 데이터입니다.
 #ifdef AFX_DESIGN_TIME
